+++ classes/class-course-blog-handler.php
@@ -110,4 +110,23 @@ class Course_Blog_Handler extends Blog_Handler {
 
 		return $wordpress_user_role;
 	}
+
+	/**
+	 * Gets the blog options to set when the blog is created or loaded
+	 *
+	 * @param array   $request_data
+	 *
+	 * @return array
+	 */
+	public function get_options_from_request( array $request_data ): array {
+		return array();
+	}
+
+	/**
+	 * Implement an abstract, but not required for course blogs
+	 *
+	 *@return void
+	 */
+	protected function fix_blog_id_from_path(): void {}
+
 }